Move example run and test
