+++ ledger-core/v2/conveyor/smachine/example/run_test.go
@@ -3,10 +3,10 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
 
-package main
+package example
 
 import "testing"
 
-func TestPlayerScenario(t *testing.T) {
-	main()
+func TestExample(t *testing.T) {
+	RunExample()
 }
